rag_snow_agent/tests/test_join_graph.py
"""Tests for JoinGraph: shortest path, confidence-aware routing."""

from __future__ import annotations

import pytest

from rag_snow_agent.retrieval.join_graph import JoinGraph, JoinGraphEdge


def _make_cards(edges: list[tuple[str, str, str, str, float, str]]) -> list[dict]:
    """Helper: build card dicts from (lt, lc, rt, rc, conf, source) tuples."""
    return [
        {
            "left_table": lt,
            "left_column": lc,
            "right_table": rt,
            "right_column": rc,
            "confidence": conf,
            "source": src,
        }
        for lt, lc, rt, rc, conf, src in edges
    ]


class TestJoinGraphBasic:
    def test_empty_graph(self):
        g = JoinGraph.from_join_cards([])
        assert g.all_tables == set()
        assert g.neighbors("X") == []

    def test_single_edge(self):
        cards = _make_cards([("A", "id", "B", "a_id", 1.0, "fk")])
        g = JoinGraph.from_join_cards(cards)
        assert g.all_tables == {"A", "B"}
        assert len(g.neighbors("A")) == 1
        assert len(g.neighbors("B")) == 1  # reverse edge

    def test_find_join_path_direct(self):
        cards = _make_cards([("A", "id", "B", "a_id", 1.0, "fk")])
        g = JoinGraph.from_join_cards(cards)
        path = g.find_join_path("A", "B")
        assert path is not None
        assert len(path) == 1
        assert path[0].right_table == "B"

    def test_find_join_path_same_table(self):
        g = JoinGraph.from_join_cards([])
        path = g.find_join_path("A", "A")
        assert path == []

    def test_find_join_path_no_connection(self):
        cards = _make_cards([("A", "id", "B", "a_id", 1.0, "fk")])
        g = JoinGraph.from_join_cards(cards)
        path = g.find_join_path("A", "C")
        assert path is None

    def test_find_join_path_two_hops(self):
        cards = _make_cards([
            ("A", "id", "B", "a_id", 1.0, "fk"),
            ("B", "id", "C", "b_id", 0.7, "heuristic_name"),
        ])
        g = JoinGraph.from_join_cards(cards)
        path = g.find_join_path("A", "C")
        assert path is not None
        assert len(path) == 2

    def test_find_join_path_respects_max_depth(self):
        cards = _make_cards([
            ("A", "id", "B", "a_id", 1.0, "fk"),
            ("B", "id", "C", "b_id", 0.7, "heuristic_name"),
            ("C", "id", "D", "c_id", 0.7, "heuristic_name"),
        ])
        g = JoinGraph.from_join_cards(cards)
        # Depth 2 should not reach D from A (needs 3 hops)
        path = g.find_join_path("A", "D", max_depth=2)
        assert path is None
        # Depth 3 should work
        path = g.find_join_path("A", "D", max_depth=3)
        assert path is not None
        assert len(path) == 3


class TestShortestBridgeTables:
    def test_no_bridge_needed_direct(self):
        cards = _make_cards([("A", "id", "B", "a_id", 1.0, "fk")])
        g = JoinGraph.from_join_cards(cards)
        bridges = g.shortest_bridge_tables(["A", "B"])
        assert bridges == []

    def test_single_bridge(self):
        cards = _make_cards([
            ("A", "id", "BRIDGE", "a_id", 1.0, "fk"),
            ("BRIDGE", "b_id", "B", "id", 0.7, "heuristic_name"),
        ])
        g = JoinGraph.from_join_cards(cards)
        bridges = g.shortest_bridge_tables(["A", "B"])
        assert "BRIDGE" in bridges

    def test_min_confidence_filter(self):
        cards = _make_cards([
            ("A", "id", "BRIDGE", "a_id", 0.3, "heuristic_name"),
            ("BRIDGE", "b_id", "B", "id", 0.3, "heuristic_name"),
        ])
        g = JoinGraph.from_join_cards(cards)
        # With min_confidence=0.5, the low-confidence path should be rejected
        bridges = g.shortest_bridge_tables(["A", "B"], min_confidence=0.5)
        assert bridges == []

    def test_single_table_no_bridge(self):
        g = JoinGraph.from_join_cards([])
        bridges = g.shortest_bridge_tables(["A"])
        assert bridges == []

    def test_prefers_shorter_path(self):
        cards = _make_cards([
            # Short path: A -> M -> B
            ("A", "id", "M", "a_id", 1.0, "fk"),
            ("M", "b_id", "B", "id", 1.0, "fk"),
            # Longer path: A -> X -> Y -> B
            ("A", "id", "X", "a_id", 1.0, "fk"),
            ("X", "y_id", "Y", "x_id", 1.0, "fk"),
            ("Y", "b_id", "B", "id", 1.0, "fk"),
        ])
        g = JoinGraph.from_join_cards(cards)
        bridges = g.shortest_bridge_tables(["A", "B"])
        # Should find M as the bridge (shorter path)
        assert "M" in bridges
        assert len(bridges) == 1


class TestConfidenceAwareRouting:
    def test_prefers_higher_confidence_same_length(self):
        """Given two equal-length paths, prefer the one with higher min confidence."""
        cards = _make_cards([
            # Path 1: A -> M1 -> B, confidence 0.7
            ("A", "id", "M1", "a_id", 0.7, "heuristic_name"),
            ("M1", "b_id", "B", "id", 0.7, "heuristic_name"),
            # Path 2: A -> M2 -> B, confidence 1.0
            ("A", "id", "M2", "a_id", 1.0, "fk"),
            ("M2", "b_id", "B", "id", 1.0, "fk"),
        ])
        g = JoinGraph.from_join_cards(cards)
        bridges = g.shortest_bridge_tables(["A", "B"])
        # BFS finds whichever first; both are length 2.
        # The confidence-aware logic should prefer M2.
        assert len(bridges) == 1
        assert bridges[0] in ("M1", "M2")  # either is valid, but M2 preferred
