rag_snow_agent/pyproject.toml
[project]
name = "rag-snow-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "scikit-learn>=1.8.0",
    "snowflake-connector-python>=4.3.0",
    "tiktoken>=0.12.0",
]
